Fix bluetooth not working bug
I was being dumb with one of the macros and a bracket was in the wrong
location so bluetooth now works

but it does seem atleast for my sega controller it shares the same
device_id as snes which might cause some problems when using bluetooth
but I haven't seen any so far. N64 works great though.